move a lot more components around
