+++ modules/client/src/routes/events/results.tsx
@@ -3,6 +3,7 @@ import {
   EventModel,
   PlayerGroupModel,
   RoundModel,
+  clamper,
 } from '@dirtleague/common';
 import { ReactElement, memo, useEffect, useState } from 'react';
 import { Grid, Label, Table, SemanticCOLORS } from 'semantic-ui-react';
@@ -28,10 +29,6 @@ const scoreLabelMap = new Map<number, string | null>([
   [1, 'orange'],
   [2, 'red'],
 ]);
-
-// TODO: Move to common.
-const clamper = (min: number, max: number) => (x: number) =>
-  Math.max(min, Math.min(x, max));
 
 const CombinedCardResults = (props: CombinedCardResultsProps): ReactElement => {
   const { model, layout } = props;
